COMP: Fix OpenCVBasicTypeBridge Matrix specialization for OpenCV 3/4
cv::Matx uses int template params; itk::Matrix uses unsigned int.
When the same non-type parameter appeared in both sides of the partial
specialization pattern, the compiler could not unify the deduction from
itk::Matrix<T, uint(2), uint(3)> (unsigned int) with cv::Matx<T, 2, 3>
(int), causing the specialization to be dropped and the empty primary
template to be selected instead.

Fix: use int for VRows/VColumns (matching OpenCV) as the deduction
variable, and place a static_cast<unsigned int> in the ITK Matrix side of the
pattern. C++17 allows integral-cast constant expressions in partial
specialization patterns, and the compiler can invert a simple integral
cast during deduction.

+++ Modules/Video/BridgeOpenCV/include/itkOpenCVBasicTypeBridge.h
@@ -161,10 +161,14 @@ struct OpenCVBasicTypeBridge<itk::Size<2>, cv::Size>
   }
 };
 
-template <typename T, unsigned int VRows, unsigned int VColumns>
-struct OpenCVBasicTypeBridge<itk::Matrix<T, VRows, VColumns>, cv::Matx<T, VRows, VColumns>>
+// cv::Matx uses int template params; itk::Matrix uses unsigned int.
+// Using int for VRows/VColumns (matching OpenCV) with a cast in the ITK
+// pattern allows the compiler to deduce both sides correctly.
+template <typename T, int VRows, int VColumns>
+struct OpenCVBasicTypeBridge<itk::Matrix<T, static_cast<unsigned int>(VRows), static_cast<unsigned int>(VColumns)>,
+                             cv::Matx<T, VRows, VColumns>>
 {
-  using ITKDataType = itk::Matrix<T, VRows, VColumns>;
+  using ITKDataType = itk::Matrix<T, static_cast<unsigned int>(VRows), static_cast<unsigned int>(VColumns)>;
   using OpenCVDataType = cv::Matx<T, VRows, VColumns>;
 
   static ITKDataType
